Move cli code out of __init__.py
It's best to keep __init__.py mostly empty and free from
core logic code
